requirements.txt
fastapi
python-dotenv
psycopg
uvicorn
